+++ scripts/run-cooler.sh
@@ -2,9 +2,9 @@
 pairs_file=$1  # bgzipped, with .px2 index
 chrsize_file=$2
 bin_size=$3
-max_iter=$4 # e.g. 3000 (default 200)
-ncores=$5 # (default 8)
-out_prefix=$6
+ncores=$4 # (default 8)
+out_prefix=$5
+
 
 cp $chrsize_file ./tmpchrsize
 
@@ -13,7 +13,4 @@ cp $chrsize_file ./tmpchrsize
 
 # the cload command requires the chrom size file to exist besides the chrom size bin file.
 cooler cload pairix -p $ncores ./tmpchrsize:$bin_size $pairs_file $out_prefix.cool
-
-# matrix balancing
-cooler balance --max-iters $max_iter $out_prefix.cool