+++ scripts/run-juicebox-pre.sh
@@ -9,12 +9,12 @@ higlass=$5   # if 1, higlass-compatible aggregation
 if [[ $higlass == '1' ]]
 then
     reslist=$(python3 -c "from cooler.contrib import higlass; higlass.print_zoom_resolutions('$chromsizefile', $min_res)")
-    java -jar juicer_tools pre -n $input_pairs $output_prefix.hic $chromsizefile -r $reslist
+    java -jar /usr/local/bin/juicer_tools.jar pre -n $input_pairs $output_prefix.hic $chromsizefile -r $reslist
 else
-    java -jar juicer_tools pre -n $input_pairs $output_prefix.hic $chromsizefile
+    java -jar /usr/local/bin/juicer_tools.jar pre -n $input_pairs $output_prefix.hic $chromsizefile
 fi
 
 # normalization
-java -jar juicer_tools addNorm -w $min_res -d -F $output_prefix.hic
+java -jar /usr/local/bin/juicer_tools.jar addNorm -w $min_res -d -F $output_prefix.hic